+++ ParentsAbroad.Interfaces/Repositories/IChildLanguageRepository.cs
@@ -5,7 +5,7 @@ namespace ParentsAbroad.Interfaces.Repositories
     public interface IChildLanguageRepository
     {
         Task<bool> AddLanguageAsync(ChildLanguage entity);
-        Task<ChildLanguage> GetAsync(long childId, long languageId);
+        Task<ChildLanguage> GetAsync(int childId, int languageId);
         Task<bool> DeleteLanguageAsync(ChildLanguage entity);
     }
 }
