+++ ParentsAbroad.Interfaces/Repositories/IChildRepository.cs
@@ -4,7 +4,7 @@ namespace ParentsAbroad.Interfaces.Repositories
 {
     public interface IChildRepository : IBaseRepository<Child>
     {
-        Task<Child> GetByIdWithRelationsAsync(long parentId);
+        Task<Child> GetByIdWithRelationsAsync(int parentId);
         Task<IList<Child>> GetAllChildrenWithRelationsAsync();
     }
 }
