+++ ParentsAbroad.Interfaces/Repositories/IChildSchoolSubjectRepository.cs
@@ -5,7 +5,7 @@ namespace ParentsAbroad.Interfaces.Repositories
     public interface IChildSchoolSubjectRepository
     {
         Task<bool> AddSchoolSubjectAsync(ChildSchoolSubject entity);
-        Task<ChildSchoolSubject> GetAsync(long childId, long schoolSubjectId);
+        Task<ChildSchoolSubject> GetAsync(int childId, int schoolSubjectId);
         Task<bool> DeleteSchoolSubjectAsync(ChildSchoolSubject entity);
     }
 }
