+++ ParentsAbroad.Interfaces/Repositories/IParentLanguageRepository.cs
@@ -5,7 +5,7 @@ namespace ParentsAbroad.Interfaces.Repositories
     public interface IParentLanguageRepository
     {
         Task<bool> AddLanguageAsync(ParentLanguage entity);
-        Task<ParentLanguage> GetAsync(long parentId, long languageId);
+        Task<ParentLanguage> GetAsync(int parentId, int languageId);
         Task<bool> DeleteLanguageAsync(ParentLanguage entity);
     }
 }
