+++ ParentsAbroad.Interfaces/Repositories/IParentRepository.cs
@@ -4,7 +4,7 @@ namespace ParentsAbroad.Interfaces.Repositories
 {
     public interface IParentRepository : IBaseRepository<Parent>
     {
-        Task<Parent> GetParentWithRelationsAsync(long parentId);
+        Task<Parent> GetParentWithRelationsAsync(int parentId);
         Task<IList<Parent>> GetAllParentsWithRelationsAsync();
     }
 }
