+++ ParentsAbroad.Interfaces/Services/IChildService.cs
@@ -11,17 +11,17 @@ namespace ParentsAbroad.Interfaces.Services
     public interface IChildService
     {
         Task<ChildDto> GetByAsync(Expression<Func<Child, bool>> filter);
-        Task<ChildDto> GetByIdAsync(long id, bool withRelations);
+        Task<ChildDto> GetByIdAsync(int id, bool withRelations);
         Task<IList<ChildDto>> GetAllAsync(bool withRelations);
-        Task<IList<ChildDto>> GetAllChildrenFromFamilyAsync(long familyId);
+        Task<IList<ChildDto>> GetAllChildrenFromFamilyAsync(int familyId);
         Task<ResponseResult<ChildDto>> AddAsync(ChildCreateUpdateDto child);
         Task<ChildDto> UpdateAsync(ChildCreateUpdateDto child);
-        Task<bool> DeleteAsync(long id);
+        Task<bool> DeleteAsync(int id);
         Task<ResponseResult<bool>> AddLanguageAsync(AddLanguageDto addLanguageDto);
-        Task<bool> DeleteLanguageAsync(long childId, long languageId);
+        Task<bool> DeleteLanguageAsync(int childId, int languageId);
         Task<ResponseResult<bool>> AddSchoolSubjectAsync(AddSchoolSubjectDto addSchoolSubjectDto);
-        Task<bool> DeleteSchoolSubjectAsync(long childId, long schoolSubjectId);
+        Task<bool> DeleteSchoolSubjectAsync(int childId, int schoolSubjectId);
         Task<ResponseResult<bool>> AddLikeToDoThingAsync(LikeToDoAddDto likeToDoAddDto);
-        Task<bool> DeleteLikeToDoThingAsync(long childId, long likeToDoId);
+        Task<bool> DeleteLikeToDoThingAsync(int childId, int likeToDoId);
     }
 }
