+++ ParentsAbroad.Interfaces/Services/IParentService.cs
@@ -10,15 +10,15 @@ namespace ParentsAbroad.Interfaces.Services
     public interface IParentService
     {
         Task<ParentDto> GetByAsync(Expression<Func<Parent, bool>> filter);
-        Task<ParentDto> GetByIdAsync(long id, bool withRelations);
+        Task<ParentDto> GetByIdAsync(int id, bool withRelations);
         Task<IList<ParentDto>> GetAllAsync(bool withRelations);
-        Task<IList<ParentDto>> GetAllParentsFromFamilyAsync(long familyId);
+        Task<IList<ParentDto>> GetAllParentsFromFamilyAsync(int familyId);
         Task<ResponseResult<ParentDto>> AddAsync(ParentCreateUpdateDto parent);
         Task<ParentDto> UpdateAsync(ParentCreateUpdateDto parent);
-        Task<bool> DeleteAsync(long id);
+        Task<bool> DeleteAsync(int id);
         Task<ResponseResult<bool>> AddLanguageAsync(AddLanguageDto addLanguageDto);
-        Task<bool> DeleteLanguageAsync(long parentId, long languageId);
+        Task<bool> DeleteLanguageAsync(int parentId, int languageId);
         Task<ResponseResult<bool>> AddHobbyAsync(AddHobbyDto addHobbyDto);
-        Task<bool> DeleteHobbyAsync(long parentId, long hobbyId);
+        Task<bool> DeleteHobbyAsync(int parentId, int hobbyId);
     }
 }
